runtime: rewrite traceMap to scale better
The existing implementation of traceMap is a hash map with a fixed
bucket table size which scales poorly with the number of elements added
to the map. After a few thousands elements are in the map, it tends to
fall over.

Furthermore, cleaning up the trace map is currently non-preemptible,
without very good reason.

This change replaces the traceMap implementation with a simple
append-only concurrent hash-trie. The data structure is incredibly
simple and does not suffer at all from the same scaling issues.

Because the traceMap no longer has a lock, and the traceRegionAlloc it
embeds is not thread-safe, we have to push that lock down. While we're
here, this change also makes the fast path for the traceRegionAlloc
lock-free. This may not be inherently faster due to contention on the
atomic add, but it creates an easy path to sharding the main allocation
buffer to reduce contention in the future. (We might want to also
consider a fully thread-local allocator that covers both string and
stack tables. The only reason a thread-local allocator isn't feasible
right now is because each of these has their own region, but we could
certainly group all them together.)

Lastly, this change also takes this opportunity to push down the
go:systemstack requirement on a whole bunch of functions. There's really
no reason it needs to be applied to much of the traceMap code and other
code in the tracer. The only thing sensitive to preemption is trace.lock
and we've already made sure most accesses happen on the system stack
right where the lock is acquired; this change just pushes the call to
systemstack closer to the actual acquisition for a bunch of additional
cases.

Change-Id: I8c06d42825c326061a1b8569e322afc4bc2a513a