cmd/link: fix bug in relro handling of .dynamic/.got sections
DO NOT SUBMIT

This patch fixes an issue introduced in CL 473495 relating to the
handling of the .dynamic/.got sections for ELF targets. The change
made in that CL was to move the contents of these sections into the
relro segment, but this was done by directly assigning the section
type, without exposing the symbols in question to the other relro
processing that happens in the linker's dodata phase.

This patch adds a new symbol type, "SELFRELROSECT", to be used to ELF
sections such as .dynamic/.got that will wind up in relro if the
target supports it, and then a call in the relro portion of dodata()
to properly assign them to the correct section.

Change-Id: I41990e7f64d8ccbc997e6013ef728c931fc33273