cmd/internal/obj/ppc64: don't modify runtime.elf_* symbols
The runtime.elf_* symbols are assembly functions which are used
to support the gcc/llvm -Os option when used with cgo.

When compiling Go for shared code, we attempt to strip out the
TOC regenation code added by the go assembler for these symbols.

This causes the symbol to no longer appear as an assembly
function which causes problems later on when handling other
implicit symbols.

Instead, account for the TOC regeneration and skip it. Note,
it does not exist if PC relative instructions are supported.

Fixes #66265

Change-Id: Icbf8e4438d177082a57bb228e39b232e7a0d7ada