cmd/internal/obj: on PPC64, don't generate stackmap symbols for linker functions
The runtime.elf_* functions are used to support linking external ELF
objects which are compiled with the gcc -Os option. They are enabled
as needed by the linker. They are never used by Go, and are only marked
reachable when the first usage is encountered. This seems to cause
issues with deadcode elimination.

These functions are written in Go asm for convenience, but referenced
by non-go symbols. Avoid adding the extra bookkeeping symbols which get
tripped up when the symbols are marked reachable.

Fixes #66265

Change-Id: I4a342d50914c9aa1c0ed03b9fbdfe534e7c35a0b