cmd/link: support -bindnow option and permit use of "-Wl,-z,now"
This is a partial roll-forward of CL https://go.dev/cl/c/go/+/473495,
which was subsequently reverted. The second half of CL 473495 will
appear in a future CL.

In this patch we introduce a new Go linker "-bindnow" command line
flag, and update the Go command to permit the use of the -Wl,-z,now
option, to allow users to produce binaries that have immediate
binding.

Updates #45681.

Change-Id: Idd61b0d6597bcd37b16c343714c55a4ef6dfb534