Fix comment typo on slog package
I'm studying the slog library for a presentation and realized this comment might be wrong.

Just a suggested correction.

It will be an honor to be able to collaborate with the project, even if it is just a drop of water. 💧

Change-Id: I3c496b5a1e30a27188a54593840225f28c79470e
GitHub-Last-Rev: 800ee2d0e8f06485914cff1e115926ff2deb5068
GitHub-Pull-Request: golang/go#66348