Fix comment typo on slog package
I'm studying the slog library for a presentation and realized this comment might be wrong.

just a suggested correction.

It will be an honor to be able to collaborate with the project, even if it is just a drop of water. 💧

Change-Id: I3ea821075522aa08694c2bbe19487ad9d402f1d7
GitHub-Last-Rev: 800ee2d0e8f06485914cff1e115926ff2deb5068
GitHub-Pull-Request: golang/go#66349