go/types: changes to support materialized aliases
This CL is a mix of a number of disparate changes (to the
compiler, go/types, and gcimporter) required to get the
tests of both go/types and x/tools working with materialized
type aliases (GODEBUG=gotypesaliases=1).

WORK IN PROGRESS: need to tease these apart into
separate CLs with tests.

See also CL XXX, which contains the corresponding x/tools
changes.

Change-Id: I70b9279f4e0ae7a1f95ad153c4e6909a878915a4