go/types: changes to support materialized aliases
This CL is a mix of a number of disparate changes (to the
compiler, go/types, and gcimporter) required to get the
tests of both go/types and x/tools working with materialized
type aliases (GODEBUG=gotypesaliases=1).

See also CL 574717, which contains the corresponding x/tools
changes.

WORK IN PROGRESS:
- need to tease these apart into separate CLs with tests.
- uses of GODEBUG need to be recorded via telemetry;
  see #66216.

Fixes #66550
Updates #65294
Updates #64581

Change-Id: I70b9279f4e0ae7a1f95ad153c4e6909a878915a4