cmd/compile, cmd/internal: fine-grained fiddling with loop alignment
This appeears to be useful only on amd64, and was specifically
benchmarked on Apple Silicon and did not produce any benefit there.
This CL adds the assembly instruction `PCALIGNIF align,amount`
which aligns to `align` if amount or fewer bytes remain. (0 means
never, but will align the enclosing function.)

Specifically, if low-order-addresss-bits + amount are
greater than or equal to align; thus, `PCALIGNIF 64,63` is
the same as `PCALIGN 64` and `PCALIGNIF 64,0` will never
emit any alignment, but will still cause the function itself
to be aligned to (at least) 64 bytes.

Change-Id: Id51a056f1672f8095e8f755e01f72836c9686aa3